+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0] - 2026-01-13
+### Added
+- New `base32` module with encoding and decoding of four variants of Base32.
+- Tests for the `base32` module.
+
 ## [2.2.2] - 2026-01-13
 ### Fixed
 - A few typos and bugs in the HMAC/HOTP/TOTP libraries I added last time.
@@ -63,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/supernovus/lum.encode.js/compare/v2.2.2...HEAD
+[Unreleased]: https://github.com/supernovus/lum.encode.js/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/supernovus/lum.encode.js/compare/v2.2.2...v2.3.0
 [2.2.2]: https://github.com/supernovus/lum.encode.js/compare/v2.2.0...v2.2.2
 [2.2.0]: https://github.com/supernovus/lum.encode.js/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/supernovus/lum.encode.js/compare/v2.0.0...v2.1.0